frontend/requirements.txt
streamlit
httpx
python-dotenv
black
flake8
isort
